compiler warning This change fixes a compiler warning on a format string.

+++ lib/srec/output/file/cosmac.cc
@@ -62,14 +62,14 @@ srec_output_file_cosmac::write(const srec_record &record)
 	if (header_required)
 	{
 	    address = record.get_address();
-	    put_stringf("!M%.*lX ", address_length, address);
+	    put_stringf("!M%.*lX ", (int)address_length, address);
 	    column = address_length + 3;
 	    header_required = false;
 	}
 	if (address != record.get_address())
 	{
 	    address = record.get_address();
-	    put_stringf(";\n%.*lX ", address_length, address);
+	    put_stringf(";\n%.*lX ", (int)address_length, address);
 	    column = address_length + 1;
 	}
 	for (int j = 0; j < record.get_length(); ++j)
